+++ code/gms-backend/src/main/java/io/github/gms/job/MessageCleanupJob.java
@@ -6,6 +6,7 @@ import io.github.gms.functions.message.MessageRepository;
 import io.github.gms.functions.systemproperty.SystemPropertyService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.core.env.Environment;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
@@ -24,13 +25,17 @@ public class MessageCleanupJob extends AbstractLimitBasedJob {
 
 	private final MessageRepository messageRepository;
 
-	public MessageCleanupJob(Clock clock, MessageRepository messageRepository, SystemPropertyService systemPropertyService) {
-		super(clock, systemPropertyService);
+	public MessageCleanupJob(Environment environment, Clock clock, MessageRepository messageRepository, SystemPropertyService systemPropertyService) {
+		super(environment, clock, systemPropertyService);
 		this.messageRepository = messageRepository;
 	}
 	
 	@Scheduled(cron = "0 0 * * * ?")
 	public void execute() {
+		if (skipJobExecution(SystemProperty.MESSAGE_CLEANUP_RUNNER_CONTAINER_ID)) {
+			return;
+		}
+
 		int result = messageRepository.deleteAllEventDateOlderThan(processConfig(SystemProperty.JOB_OLD_MESSAGE_LIMIT));
 
 		if (result > 0) {
